fix: PostBouquetAPI 꽃 명세서 + 요구사항 이미지 통합해서 보낼 수 있도록 수정
